% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_evaluation.R
\name{calculate_school_stats}
\alias{calculate_school_stats}
\title{Berechnet die statistische uebersichtstabelle der abgebenden Schulen}
\usage{
calculate_school_stats(df, all_schools_leverage, must_link_schools_vec)
}
\arguments{
\item{df}{Ein Dataframe mit den finalen Schuelerdaten.}

\item{all_schools_leverage}{Dataframe mit max_leverage aller Schulen.}

\item{must_link_schools_vec}{Vektor der Schulen als Must-Link gelten.}
}
\value{
Ein Dataframe \code{school_stats}.
}
\description{
Erzeugt eine aggregierte uebersichtstabelle fuer alle Schulen, inklusive Leverage.
}
