% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_clean_data.R
\name{read_and_clean_data}
\alias{read_and_clean_data}
\title{Liest die Klassenzusammensetzungs-Excel ein und bereitet sie vor.}
\usage{
read_and_clean_data(file_path)
}
\arguments{
\item{file_path}{Ein String. Der absolute oder relative Pfad zur Excel-Datei,
die die Schuelerdaten enthaelt.}
}
\value{
Ein Dataframe (\code{tibble}) mit den bereinigten und typisierten Daten.
Die wichtigsten Spalten:
\describe{
\item{name, vorname}{Name und Vorname des Kindes.}
\item{geschlecht}{Geschlecht ("m" oder "w").}
\item{abgebende_schule}{Der Name der Grundschule.}
\item{dg}{Durchschnittsnote gewichtet.}
\item{ds}{Durchschnittsnote Sprache.}
\item{de}{Durchschnittsnote einfach.}
\item{ue}{uebergangsempfehlung (numerisch kodiert 1 bis 5).}
\item{mig}{Migrationshintergrund ("ja" oder "nein").}
}
}
\description{
Importiert die Rohdaten aus einer Excel-Datei, bereinigt die Spaltennamen
und fuehrt erste Datentransformationen durch. Dies umfasst das Umbenennen,
Typisieren und Filtern (z.B. Entfernung von Schuelern, die an andere Schulen
weiterziehen).
}
